+++ src/app/modules/marketplace/components/modules-detail/ai4eosc-module-detail/ai4eosc-module-detail.component.spec.ts
@@ -36,6 +36,7 @@ const mockedModule: Ai4eoscModule = {
     categories: ['cat1'],
     tags: ['tag1'],
     id: 'test',
+    license: 'MIT',
 };
 
 const mockedMediaQueryList: MediaQueryList = {