+++ src/app/modules/marketplace/components/modules-list/modules-list.component.ts
@@ -7,11 +7,13 @@ import {
     Ai4lifeModule,
     ModuleSummary,
 } from '@app/shared/interfaces/module.interface';
-import { forkJoin } from 'rxjs';
+import { filter, forkJoin } from 'rxjs';
 import { ToolsService } from '../../services/tools-service/tools.service';
 import { MatTabChangeEvent, MatTabGroup } from '@angular/material/tabs';
 import { SnackbarService } from '@app/shared/services/snackbar/snackbar.service';
 import { IntroJSService } from 'introjs/introjs.service';
+import { AppConfigService } from '@app/core/services/app-config/app-config.service';
+import { NavigationEnd, Router } from '@angular/router';
 
 @Component({
     selector: 'app-modules-list',
@@ -20,17 +22,36 @@ import { IntroJSService } from 'introjs/introjs.service';
 })
 export class ModulesListComponent implements OnInit {
     constructor(
+        private router: Router,
         private media: MediaMatcher,
         private changeDetectorRef: ChangeDetectorRef,
         private modulesService: ModulesService,
         private toolsService: ToolsService,
         private snackbarService: SnackbarService,
         private introService: IntroJSService,
+        private appConfigService: AppConfigService,
         public dialog: MatDialog
     ) {
         this.mobileQuery = this.media.matchMedia('(max-width: 600px)');
         this._mobileQueryListener = () => changeDetectorRef.detectChanges();
         this.mobileQuery.addEventListener('change', this._mobileQueryListener);
+
+        // scroll to last scrollY position
+        this.router.events
+            .pipe(filter((event) => event instanceof NavigationEnd))
+            .subscribe(() => {
+                const scrollTop = sessionStorage.getItem('scrollTop');
+                if (scrollTop) {
+                    setTimeout(() => {
+                        const content = document.querySelector(
+                            '.sidenav-content'
+                        ) as HTMLElement;
+                        if (content) {
+                            content.scrollTop = +scrollTop;
+                        }
+                    }, 100);
+                }
+            });
     }
 
     @ViewChild('tabGroup', { static: false }) tabGroup!: MatTabGroup;
@@ -75,9 +96,13 @@ export class ModulesListComponent implements OnInit {
 
     ngAfterViewInit(): void {
         const interval = setInterval(() => {
-            if (!this.ai4eoscModulesLoading && !this.ai4lifeModulesLoading) {
+            if (
+                !this.ai4eoscModulesLoading &&
+                !this.ai4lifeModulesLoading &&
+                this.appConfigService.voName !== 'vo.imagine-ai.eu'
+            ) {
                 clearInterval(interval);
-                this.introService.ai4lifeMarketplace();
+                this.introService.llmTool();
             }
         }, 200);
     }