+++ src/app/shared/components/deployments-table/deployments-table.component.ts
@@ -32,6 +32,8 @@ import {
 } from '@app/modules/deployments/utils/deployment-badge';
 import { Router } from '@angular/router';
 import { SnapshotDetailComponent } from '@app/modules/deployments/components/snapshot-detail/snapshot-detail.component';
+import { StatusNotification } from '@app/shared/interfaces/platform-status.interface';
+import { TranslateService } from '@ngx-translate/core';
 
 @Component({
     selector: 'app-deployments-table',
@@ -43,6 +45,7 @@ export class DeploymentsTableComponent implements OnInit, OnDestroy {
         public dialog: MatDialog,
         private snackbarService: SnackbarService,
         private snapshotService: SnapshotService,
+        public translateService: TranslateService,
         public confirmationDialog: MatDialog,
         private media: MediaMatcher,
         private router: Router,
@@ -63,6 +66,7 @@ export class DeploymentsTableComponent implements OnInit, OnDestroy {
     @Input() isLoading = false;
     @Input() dataset: Array<DeploymentTableRow> = [];
     @Input() dataSource!: MatTableDataSource<DeploymentTableRow>;
+    @Input() datacentersNotifications: StatusNotification[] = [];
 
     @Output() showElementInfo = new EventEmitter<string>();
     @Output() deleteElement = new EventEmitter<string>();
@@ -76,6 +80,7 @@ export class DeploymentsTableComponent implements OnInit, OnDestroy {
         { columnDef: 'creationTime', header: 'DEPLOYMENTS.CREATION-TIME' },
         { columnDef: 'endpoints', header: '', hidden: true },
         { columnDef: 'description', header: '', hidden: true },
+        { columnDef: 'datacenter', header: '', hidden: true },
         { columnDef: 'actions', header: 'DEPLOYMENTS.ACTIONS' },
     ];
 
@@ -103,6 +108,7 @@ export class DeploymentsTableComponent implements OnInit, OnDestroy {
                 size: +row.size!,
                 nomad_ID: '',
                 description: row.description,
+                error_msg: row.error_msg,
             };
             this.openSnapshotDetailDialog(snapshot);
         } else {
@@ -156,7 +162,7 @@ export class DeploymentsTableComponent implements OnInit, OnDestroy {
     }
 
     isFederatedServer(row: DeploymentTableRow) {
-        return row.mainEndpoint?.includes('fedserver');
+        return row.containerName?.includes('ai4os-federated-server');
     }
 
     getDeploymentEndpoints(row: DeploymentTableRow) {
@@ -178,6 +184,39 @@ export class DeploymentsTableComponent implements OnInit, OnDestroy {
 
     hasDeploymentErrors(row: DeploymentTableRow) {
         return row.error_msg;
+    }
+
+    hasDatacenterUnderMaintenance(row: DeploymentTableRow) {
+        return (
+            row.datacenter &&
+            this.datacentersNotifications.find((n) =>
+                n.datacenters?.includes(row.datacenter!)
+            )
+        );
+    }
+
+    getMaintenanceInfo(row: DeploymentTableRow): string {
+        let info = '';
+        const datacenterNotification = this.datacentersNotifications.find((n) =>
+            n.datacenters?.includes(row.datacenter!)
+        );
+        if (datacenterNotification) {
+            info = this.translateService.instant(
+                'DEPLOYMENTS.DATACENTER-DOWNTIME-NOTIFICATION',
+                {
+                    datacenter: row.datacenter,
+                    startDate:
+                        datacenterNotification.downtimeStart?.toLocaleDateString(
+                            'es-ES'
+                        ),
+                    endDate:
+                        datacenterNotification.downtimeEnd?.toLocaleDateString(
+                            'es-ES'
+                        ),
+                }
+            );
+        }
+        return info;
     }
 
     returnDeploymentBadge(status: string) {
@@ -211,11 +250,8 @@ export class DeploymentsTableComponent implements OnInit, OnDestroy {
                                 );
                             }
                         },
-                        error: () => {
-                            this.snackbarService.openError(
-                                'Error creating snapshot of deployment with uuid: ' +
-                                    row.uuid
-                            );
+                        error: (error) => {
+                            this.snackbarService.openError(error);
                         },
                     });
                 }